Added support for different compression types, namely 7Zip
Still able to use the older zip method, but now can use 7Zip. It will
launch 7Z command line and add the file into it, and place it in the
system drive root just like before. Since it uses 7z it does REQUIRE 7Z
to be installed. We do some checking first to ensure this is the case,
and cater for different install locations.
Also put the textwriter's initiation & using statement INSIDE the
ExportStart() method so we can properly dispose of it. It should have
never been outside.